requirements.txt
Flask==1.1.4
lark==1.1.1
gunicorn==19.9.0
flask-compress==1.4.0
requests==2.23.0
attrs==19.3.0
Flask-Commonmark==0.8
bcrypt==3.2.0
boto3>=1.16.50
MarkupSafe==2.0.1
ruamel.yaml==0.17.4
docopt==0.6.2
pydantic==1.8.2
lazy==1.4
pylint==2.8.2
awscli>=1.19.88
PySumTypes==0.0.1
beautifulsoup4==4.9.3
regex==2021.8.28
retrying==1.3.3
pytest==6.2.5
parameterized==0.8.1
Flask-Babel==2.0.0
iso3166~=2.0.2
turtlethread>=0.0.6
PyQt6-WebEngine~=6.3